requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
Pillow>=9.5.0 